refactor: enum class 디렉토리 수정
