--- src/main/java/com/kurierfree/server/domain/user/Gender.java
@@ -1,5 +1,0 @@
-package com.kurierfree.server.domain.user;
-
-public enum Gender {
-    Male, Female
-}
