+++ src/main/java/com/kurierfree/server/domain/user/domain/Gender.java
@@ -1,0 +1,5 @@
+package com.kurierfree.server.domain.user.domain;
+
+public enum Gender {
+    Male, Female
+}
