+++ src/main/java/com/kurierfree/server/domain/user/domain/Role.java
@@ -1,4 +1,4 @@
-package com.kurierfree.server.domain.user;
+package com.kurierfree.server.domain.user.domain;
 
 public enum Role {
     SUPPORTER, DISABLED_STUDENT
